added function to download keras forecast
